Following a different approach to retrieve all the necessary information
+ Added a ProjectProcessor class that retrieves all classes' names and
packages' names, and transforms them into fully qualified names (to be
able to use reflection's forName() method).

UMLDiagramsGenerator class will now rely on this class to generate the
diagrams, in the following way :
--> It draws a package diagram and adds any classes associated with it
to it (by checking its package via getPackageName()) whenever it
encounters a new package name.

+ Enhanced the user interface :
	- added some touches to the package diagram to make it more readable.
	- created a ClassDiagram class that is going to be representing class
diagrams.

The approach we're working towards to draw the diagrams is the following
:
- Each row represents a package.
- Each package contains the classes it has.
- Between each row, we leave some space for relations of classes.
- We lave some space in the right side of these rows for package
relations.